// packages/viewer/src/store/modules/ui/types/featureInfoPositions.enum.ts
export enum FeatureInfoPositions {
    /**
     * This is not the default value, but this is the default behavior,
     *     which depends on the UI size. Bottompanel on phones, tooltip on desktop
     */
    Default = 'default',
    BottomPanel = 'bottomPanel',
    ToolTip = 'tooltip',
    None = 'none',
}